PLANET-4802 Remove twig templates and use JSX instead
