+++ assets/src/blocks/Columns/ColumnTypes.js
@@ -1,0 +1,4 @@
+export const LAYOUT_NO_IMAGE = 'no_image';
+export const LAYOUT_TASKS    = 'tasks';
+export const LAYOUT_ICONS    = 'icons';
+export const LAYOUT_IMAGES   = 'image';
