--- templates/blocks/columns.twig
@@ -1,81 +1,0 @@
-{% set is_campaign = ( 'campaign' == fn( 'get_post_type' ) ) %}
-
-{% block columns %}
-	{% if ( fields ) %}
-		{% if fields.columns_block_style == 'tasks' %}
-			{% include 'columns_tasks.twig' %}
-		{% else %}
-			<section class="block columns-block block-style-{{ fields.columns_block_style }}">
-				<div class="container">
-					{% if ( fields.columns_title ) %}
-						<header>
-							<h2 class="page-section-header">{{ fields.columns_title|e('wp_kses_post')|raw }}</h2>
-						</header>
-					{% endif %}
-					{% if ( fields.columns_description ) %}
-						<div class="page-section-description">{{ fields.columns_description|e('wp_kses_post')|wpautop|raw }}</div>
-					{% endif %}
-
-					<div class="row">
-						{% for col in fields.columns %}
-							<div class="col-md-6 col-lg column-wrap">
-								{% if col.attachment and fields.columns_block_style != 'no_image' %}
-									<div class="attachment-container">
-										{% if col.cta_link %}
-											<a
-												href="{{ col.cta_link|e('esc_url') }}"
-												{{  col.link_new_tab and col.link_new_tab != "false" ? 'target="_blank"' }}
-												data-ga-category="Columns Block"
-												data-ga-action="{{ fields.columns_block_style == 'icons' ? 'Icon' : 'Image' }}"
-												data-ga-label="{{ col.cta_link|e('esc_url') }}">
-										{% endif %}
-											<img src="{{ col.attachment }}" alt="{{ col.title }}">
-										{% if col.cta_link %}
-											</a>
-										{% endif %}
-									</div>
-								{% endif %}
-
-								{% if col.cta_link and is_campaign %}
-									<h3
-										data-ga-category="Columns Block"
-										data-ga-action="Title"
-										data-ga-label="n/a">
-								{% elseif col.cta_link %}
-									<h3>
-										<a
-											href="{{ col.cta_link|e('esc_url') }}"
-											data-ga-category="Columns Block"
-											data-ga-action="Title"
-											data-ga-label="{{ col.cta_link|e('esc_url') }}">
-								{% else %}
-									<h3>
-								{% endif %}
-											{{ col.title|e('wp_kses_post')|raw }}
-								{% if col.cta_link and not is_campaign %}</a>{% endif %}
-									</h3>
-								{% if col.description %}
-									<p>{{ col.description|e('wp_kses_post')|raw|nl2br }}</p>
-								{% endif %}
-								{% if col.cta_text and col.cta_link %}
-									<a href="{{ col.cta_link|e('esc_url') }}"
-										{{ col.link_new_tab ? 'target="_blank"' }}
-										{% if is_campaign or fields.columns_block_style == 'no_image' %}
-											class="btn {{ is_campaign ? 'btn-primary' : 'btn-secondary' }}"
-										{% else %}
-											class="call-to-action-link"
-										{% endif %}
-											data-ga-category="Columns Block"
-											data-ga-action="Call to Action"
-											data-ga-label="{{ col.cta_link|e('esc_url') }}">
-										{{ col.cta_text|e('wp_kses_post')|raw }}
-									</a>
-								{% endif %}
-							</div>
-						{% endfor %}
-					</div>
-				</div>
-			</section>
-		{% endif %}
-	{% endif %}
-{% endblock %}
