--- templates/blocks/columns_tasks.twig
@@ -1,154 +1,0 @@
-{% block tasks %}
-
-	{% set is_campaign = ( 'campaign' == fn( 'get_post_type' ) ) %}
-
-	{% if ( fields ) %}
-
-		<section class="block columns-block block-style-{{ fields.columns_block_style }}">
-			<div class="container">
-				{% if ( fields['columns_title'] ) %}
-					<header>
-						<h2 class="page-section-header">{{ fields['columns_title'] }}</h2>
-					</header>
-				{% endif %}
-				{% if ( fields['columns_description'] ) %}
-					<div class="page-section-description">{{ fields['columns_description']|e('wp_kses_post')|raw }}</div>
-				{% endif %}
-
-				<div class="tasks-wrap can-do-steps d-none d-lg-block">
-					<div class="row">
-						{% for col in fields.columns %}
-							<div class="col-md-6 col-lg column-wrap step-info">
-								<span class="step-number">
-									<span
-										class="step-number-inner"
-										data-ga-category="Columns Block"
-										data-ga-action="Task Number"
-										data-ga-label="n/a">
-											{{ loop.index }}
-									</span>
-								</span>
-								<h5>
-									{% if ( col.cta_link ) %}
-										<a
-											href="{{ col.cta_link|e('esc_url') }}"
-											{{  col.link_new_tab  ? 'target="_blank"' }}
-											data-ga-category="Columns Block"
-											data-ga-action="Title"
-											data-ga-label="{{ col.cta_link|e('esc_url') }}"
-										>
-									{% endif %}
-										{{ col.title }}
-									{% if ( col.cta_link ) %}
-										</a>
-									{% endif %}
-								</h5>
-								{% if ( col.description ) %}
-									<p>
-										{{ col.description|e('wp_kses_post')|raw }}
-									</p>
-								{% endif %}
-								{% if ( col.attachment ) %}
-									<img src="{{ col.attachment }}" alt=""/>
-								{% endif %}
-								{% if ( col.cta_text and col.cta_link ) %}
-									<a
-										class="btn btn-small {{ is_campaign ? 'btn-primary' : 'btn-secondary' }}"
-										href="{{ col.cta_link|e('esc_url') }}"
-										{{  col.link_new_tab  ? 'target="_blank"' }}
-										data-ga-category="Columns Block"
-										data-ga-action="Call to Action"
-										data-ga-label="{{ col.cta_link|e('esc_url') }}">
-											{{ col.cta_text }}
-									</a>
-								{% endif %}
-							</div>
-						{% endfor %}
-					</div>
-				</div>
-
-				<div class="tasks-wrap can-do-steps-mobile d-lg-none">
-					<div id="accordion" class="card-accordion" role="tablist" aria-multiselectable="true">
-						{% for col in fields.columns %}
-								{% if loop.index == 1 %}
-									{% set number_to_word = 'one' %}
-								{% elseif loop.index == 2 %}
-									{% set number_to_word = 'two' %}
-								{% elseif loop.index == 3 %}
-									{% set number_to_word = 'three' %}
-								{% elseif loop.index == 4 %}
-									{% set number_to_word = 'four' %}
-								{% endif %}
-
-								<div class="card">
-									<a class="card-header {{ loop.index>1 ? 'collapsed' : '' }}" role="tab" id="heading-{{ number_to_word }}"
-									   data-bs-toggle="collapse" data-bs-target=".card-header:hover + #collapse-{{ number_to_word }}"
-									   href="#collapse-{{ number_to_word }}"
-									   aria-expanded="true"
-									   aria-controls="collapse-{{ number_to_word }}"
-									   data-ga-category="Columns Block"
-									   data-ga-action="Title"
-									   data-ga-label="n/a">
-										<span
-											class="step-number"
-											data-ga-category="Columns Block"
-											data-ga-action="Task Number"
-											data-ga-label="n/a">
-												{{ loop.index }}
-										</span>
-										{% if ( col.title ) %}
-											{{ col.title }}
-										{% endif %}
-									</a>
-
-									{% if fields.no_of_columns <= 2 %}
-										{% set collapse_class = 'collapse show' %}
-									{% else %}
-										{% if loop.index == 1 %}
-											{% set collapse_class = 'collapse show' %}
-										{% else %}
-											{% set collapse_class = 'collapse' %}
-										{% endif %}
-									{% endif %}
-
-									<div id="collapse-{{ number_to_word }}" class="{{ collapse_class }}"
-										 data-bs-parent="#accordion" role="tabpanel"
-										 aria-labelledby="heading-{{ number_to_word }}">
-										<div class="card-block info-with-image-wrap clearfix">
-											<div class="mobile-accordion-info">
-												{% if ( col.description ) %}
-													<p>
-														{{ col.description|e('wp_kses_post')|raw }}
-													</p>
-												{% endif %}
-											</div>
-
-											{% if ( col.attachment ) %}
-												<img src="{{ col.attachment }}" alt="">
-											{% endif %}
-
-											{% if ( col.cta_text and col.cta_link ) %}
-												<a
-													class="btn btn-small {{ is_campaign ? 'btn-primary' : 'btn-secondary' }}"
-													href="{{ col.cta_link|e('esc_url') }}"
-													data-ga-category="Columns Block"
-													data-ga-action="Call to Action"
-													data-ga-label="{{ col.cta_link|e('esc_url') }}"
-													{{  col.link_new_tab  ? 'target="_blank"' }}
-												>
-														{{ col.cta_text }}
-												</a>
-											{% endif %}
-
-										</div>
-									</div>
-								</div>
-						{% endfor %}
-					</div>
-				</div>
-			</div>
-		</section>
-
-	{% endif %}
-
-{% endblock %}
